// src/types.ts
export interface Profile {
  username: string;
  bio: string;
  image: string;
  following: boolean;
}

export interface User {
  email: string;
  token: string;
  username: string;
  bio: string;
  image: string;
}

export interface Article {
  slug: string;
  title: string;
  description: string;
  body: string;
  tagList: string[];
  createdAt: string;
  updatedAt: string;
  favorited: boolean;
  favoritesCount: number;
  author: Profile;
}

export interface ApiError {
  errors: Record<string, string[]>;
}